Publisher archive added.
Each publisher that's added through the publis site is added into
archive table that is not modified afterwards. In this way it is
possible to see the original information that was received even if the
record has been modified afterwards. Records that are create through the
backend are not added into archive.