Updated issn creation process.
Now it's possible to cancel ISSN number even if it's not the last one
that was created. In addition, all the canceled ISSN numbers can be
reused, even if the range in which they belong is not active anymore.
When ISSN is canceled, the canceled ISSN number is used the next time
when new ISSN is created. If multiple ISSN numbers are canceled before a
new one is created, they will stay in a queue, and they're used starting
from the smallest one.