Fixed Joomla related bug.
Special care needs to be taken with saving an unchecked checkbox from a
form. On saving a form with a checkbox that is unchecked, there is no
variable for it in the POST data.

Solution: added  invisible hidden input field before the checkbox (with
the same name and the empty checkbox value), which the checkbox will
override if it's checked.