Improved deleting publisher identifiers
Publisher identifiers can now be deleted even if the identifier to be
deleted is not the last identifier generated from its range. On
identifier generation canceled identifiers are always used first if
there are canceled identifiers from the requested category. Canceled
identifiers from the same range are used first, and then canceled
identifiers from other ranges with the same category.