Updated ISSN next pointer handling
Earlier next pointer value was always increased when new ISSN was
generated, also when the last ISSN of the range was used. When range was
closed, the next pointer value was range_end + 1. This behaviour causes
an error when range_end is 999X. The next value would be 1000, but next
pointer length is 3 digits + check digits. Now next pointer is set to
empty when the range is closed.